Change: FS Daemon backend to Watchdog
After some review of the PyInotify and Watchdog libraries is
looked as though Watchdog had a lower bar for entry. That is, it
appeared to have a simpler interface and exposed more of the
desired functionality by default.

Should this perception prove to be incorrect, the
PyInotify library will be brought back.

Signed-off-by: Kevin Conway <kevinjacobconway@gmail.com>